Initial commit. This is an operational example of a command that takes a number of subcommands. Of particular importance is how `parse_args` is defined (see climath.py).
The individual commands, e.g. climath-add or climath-mean, are
variants on this that allows one to call each of these as
stand-alone programs. (Because all reduction operations share
a common help and option format see reduction_operation.py to
see how all reduction operations can use a generator to
customize the one (or two) things needed to produce specific
operations. (An analogous method is present in the `binary`
module, see `binary_operation.py`.)